services/order/pyproject.toml
[tool.poetry]
name = "order-service"
version = "0.1.0"
description = "FastAPI Order Service with RabbitMQ and PostgreSQL"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
pydantic = "^1.8.2"
tortoise-orm = "^0.17.7"
aerich = "^0.6.1"
asyncpg = "^0.25.0"
psycopg2-binary = "^2.9.1"
python-dotenv = "^0.19.0"
pika = "^1.2.0"
python-multipart = "^0.0.5"
email-validator = "^1.1.3"
ujson = "^4.0.2"
orjson = "^3.6.0"
aiofiles = "^0.7.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-asyncio = "^0.15.1"
httpx = "^0.18.2"
black = "^21.7b0"
isort = "^5.9.3"
flake8 = "^3.9.2"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "config.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "."

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88